example_project/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-adminlte2-templates==0.0.4
django-crispy-forms==1.9.0
pytz==2019.3
sqlparse==0.3.1
